+++ alive_progress/tools/sampling.py
@@ -1,4 +1,4 @@
-from types import SimpleNamespace
+import timeit
 
 from about_time import duration_human
 
@@ -11,15 +11,12 @@ def overhead(total=None, *, calibrate=None, **options):
     number = 400  # timeit number of runs inside each repetition.
     repeat = 300  # timeit how many times to repeat the whole test.
 
-    import timeit
-    config, sampler = config_handler(force_tty=False, **options), SimpleNamespace()
-    with __alive_bar(config, total, calibrate=calibrate, _write=__noop_p, _flush=__noop,
-                     _cond=__lock, _term_cols=__noop_z, _hook_manager=__hook_manager,
-                     _sampler=sampler) as bar:
-        sampler.__dict__.update(bar.__dict__)
+    config = config_handler(disable=True, **options)
+    with __alive_bar(config, total, calibrate=calibrate, _cond=__lock):
         # the timing of the print_cells function increases proportionately with the
-        # number of columns in the terminal, so I want a baseline here with `_term_cols=0`.
-        res = timeit.repeat('_alive_repr()', repeat=repeat, number=number, globals=sampler.__dict__)
+        # number of columns in the terminal, so I want a baseline here `VOID.cols == 0`.
+        res = timeit.repeat('_alive_repr()', repeat=repeat, number=number,
+                            globals=__alive_bar.__dict__)
 
     return duration_human(min(res) / number).replace('us', 'µs')
 
@@ -60,15 +57,7 @@ def overhead_sampling():
         print('|')
 
 
-def __noop():
-    pass
-
-
 def __noop_p(_ignore):
-    return 0
-
-
-def __noop_z():
     return 0
 
 
@@ -78,13 +67,6 @@ class __lock:
 
     def __exit__(self, _type, value, traceback):
         pass
-
-
-def __hook_manager(_=None, __=None, ___=None):
-    __hook_manager.flush_buffers = __noop
-    __hook_manager.install = __noop
-    __hook_manager.uninstall = __noop
-    return __hook_manager
 
 
 if __name__ == '__main__':